+++ packages/rust-bindings/CHANGELOG.md
@@ -5,6 +5,16 @@
 ### Added
 
 - `create_web3_id_proof_ext` function.
+- `verify_web3_id_credential_signature_ext` function.
+
+
+### Changed
+
+- Add issuer contract index and subindex as arguments to `verify_web3_id_credential_signature_ext` and `get_verifiable_credential_public_key_ext`.
+
+### Removed
+
+- `get_verifiable_credential_encryption_key_ext` function.
 
 ## 1.1.0